Fix: a few bugs in assignment, initialization, ...
Introducing some C++ unit tests uncovered a few bugs:
  * miscalculating some buffer sizes;
  * miscalculating round-up-to-power-of-two logic;
  * some copying and assignment bugs.

This change lightly refactors to make more of the above
testable, tests them, and fixes the problems that surfaced.
Code inspection also shows some memory leaks and other issues,
but Rome wasn't built in a day and the patch is getting big
as-is.