Added: Macro SZ_NULL_CHAR, Clang-CL instrinsics.
Despite declaring it as extern "C", MSVC and Clang-Cl both still enforce that you cannot convert a void*0 (r-value) to a char* for some reason. As such had to added SZ_NULL_CHAR.

Clang-Cl also picks and chooses when it wants to use MSVC intrinsics over clang instrinsics so added resolution for that.

Also build script was just added PIC even tho it might not be available.

NOTE: None of this code is tested as, both Clang-Cl and MSVC are not able to build any of the tests on my machine despite best efforts.