Improve: `sz_find_neon` for different lengths
This commit accelerates normal-order substring search on Aarch64,
by avoiding unnecessary comparisons when dealing with strings of
length 2 and 3.

A straightforward implementation yields a noticeable performance
improvement up to 20% on ASCII datasets with English texts. The
cleaner and "smarter" implementation, that avoids unnecessary loads
doesn't perform better and is placed into the "experimental.h" file
for future research.